Refactored to use Redux and moved things around
Moved to redux because there will be some sort of message based
synchronization later and state will be manipulated from a lot of places.

Don't know if I'll keep this folder hierarchy, but for now it seems ok.

Aside: I really, really like the redux documentation, and my code is a lot cleaner because of it.